Live TV select channel from another video source
In Live TV mode it is possible to select any other channel by entering the channel number.
In fixes/32 and in master this is broken; it is now only possible to select a channel in
the same video source as the current channel.
Fixed by removing a static_cast in two lines in function get_chanid that was introduced
in  commit 976989051502f54abb7173f9e9127f25f8fd6c13 on September 29, 2020.

(cherry picked from commit 61b82f2b5a4124c4369fa9ae1b408b6d36251e05)

+++ mythtv/libs/libmythtv/tv_play.cpp
@@ -5956,19 +5956,20 @@ static uint get_chanid(const PlayerContext *ctx,
 {
     uint chanid = 0;
     uint cur_sourceid = 0;
+
     // try to find channel on current input
     if (ctx && ctx->m_playingInfo && ctx->m_playingInfo->GetSourceID())
     {
         cur_sourceid = ctx->m_playingInfo->GetSourceID();
-        chanid = std::max(static_cast<uint>(ChannelUtil::GetChanID(cur_sourceid, channum)), 0U);
+        chanid = std::max(ChannelUtil::GetChanID(cur_sourceid, channum), 0);
         if (chanid)
             return chanid;
     }
-    // try to find channel on specified input
 
+    // try to find channel on specified input
     uint sourceid = CardUtil::GetSourceID(cardid);
     if (cur_sourceid != sourceid && sourceid)
-        chanid = std::max(static_cast<uint>(ChannelUtil::GetChanID(sourceid, channum)), 0U);
+        chanid = std::max(ChannelUtil::GetChanID(sourceid, channum), 0);
     return chanid;
 }
